refactor: rename ecc.rs to x25519.rs
for better clarity


tutadb#2026
