+++ tuta-sdk/rust/sdk/src/crypto/key_encryption.rs
@@ -1,8 +1,8 @@
-use crate::crypto::ecc::{EccKeyPair, EccPrivateKey, EccPublicKey};
 use crate::crypto::key::{AsymmetricKeyPair, GenericAesKey, KeyLoadError};
 use crate::crypto::kyber::{KyberKeyPair, KyberPrivateKey, KyberPublicKey};
 use crate::crypto::rsa::{RSAEccKeyPair, RSAKeyPair, RSAPrivateKey, RSAPublicKey};
 use crate::crypto::tuta_crypt::PQKeyPairs;
+use crate::crypto::x25519::{EccKeyPair, EccPrivateKey, EccPublicKey};
 use crate::entities::generated::sys::KeyPair;
 use crate::ApiCallError;
 use zeroize::Zeroizing;
@@ -112,9 +112,9 @@ fn decrypt_rsa_or_rsa_ecc_key_pair(
 #[cfg(test)]
 mod tests {
 	use super::*;
-	use crate::crypto::ecc::EccKeyPair;
 	use crate::crypto::key::{AsymmetricKeyPair, GenericAesKey};
 	use crate::crypto::randomizer_facade::test_util::make_thread_rng_facade;
+	use crate::crypto::x25519::EccKeyPair;
 	use crate::crypto::{aes::Iv, Aes256Key, PQKeyPairs};
 	use crate::entities::generated::sys::KeyPair;
 	use crate::util::test_utils::generate_random_string;
